+++ example/start.js
@@ -1,0 +1,3 @@
+// @flow
+
+console.log('Example application started!');