+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.3.15] - Unreleased
+## [0.4.0] - Unreleased
+
+### Added
+
+- Will now install and update `pm2` on the target server
+- Now supports a new `name` configuration option that will set the name of the `pm2` application (by default will use the name in the `package.json` of the app)
 
 ### Fixed